+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.2.2
 - Remove unneeded BWGPumpkin Class
+- Use Simpler Code for ManOWar Texture Location
 
 # 2.2.1
 - Match Pale jack o’lantern Light Level to Soul Fruit block light level